feat(externalEnv): add external env capability
