+++ src/cardPage.py
@@ -1,0 +1,11 @@
+import requests
+import os
+import pathlib
+import pandas
+import json
+import argparse
+import sys
+
+
+def card_page(card_name):
+    print("todo")
